<refactor>: Miniatures.js переименован в gallery-mini.js
