+++ js/main.js
@@ -1,2 +1,2 @@
-import './miniatures.js';
+import './gallery-mini.js';
 import './gallery-fullsize.js';
